refactor(web): inline simple single-use hooks
