--- apps/web/src/app/(home)/_ui/NewsSection/_hooks/useSectionHadnler.ts
@@ -1,38 +1,0 @@
-import { type RefObject, useEffect, useRef, useState } from "react";
-
-interface UseSectionHandlerReturn {
-  sectionRef: RefObject<HTMLDivElement>;
-  visible: boolean;
-}
-
-const useSectionHandler = (): UseSectionHandlerReturn => {
-  const [visible, setVisible] = useState<boolean>(false);
-
-  const sectionRef = useRef<HTMLDivElement>(null!);
-  useEffect(() => {
-    if (!sectionRef.current) return;
-
-    const observer = new window.IntersectionObserver(
-      ([entry]) => {
-        if (entry.isIntersecting) {
-          setVisible(true);
-          observer.disconnect();
-        }
-      },
-      {
-        rootMargin: "0px",
-        threshold: 0,
-      },
-    );
-
-    observer.observe(sectionRef.current);
-
-    return () => observer.disconnect();
-  }, []);
-  return {
-    sectionRef,
-    visible,
-  };
-};
-
-export default useSectionHandler;
