--- apps/web/src/app/mentor/chat/[chatId]/_ui/ChatContent/_hooks/usePutChatReadHandler.ts
@@ -1,14 +1,0 @@
-import { useEffect } from "react";
-
-import { usePutChatRead } from "@/apis/chat";
-
-const usePutChatReadHandler = (chatId: number) => {
-  const { mutate: putChatRead } = usePutChatRead();
-
-  useEffect(() => {
-    if (chatId) {
-      putChatRead(chatId);
-    }
-  }, [chatId, putChatRead]);
-};
-export default usePutChatReadHandler;
