+++ apps/web/src/components/mentor/MentorExpandChatCard/index.tsx
@@ -1,10 +1,11 @@
 "use client";
 
 import Link from "next/link";
-import usePatchApprovalStatusHandler from "@/components/mentor/MentorExpandChatCard/hooks/usePatchApprovalStatusHandler";
+import { usePatchApprovalStatus } from "@/apis/mentor";
 import ProfileWithBadge from "@/components/ui/ProfileWithBadge";
-import { IconDirectionDown, IconDirectionUp } from "@/public/svgs/mentor";
-import { MentoringApplyStatus } from "@/types/mentor";
+import { customConfirm } from "@/lib/zustand/useConfirmModalStore";
+import { IconDirectionDown, IconDirectionUp, IconUnSmile } from "@/public/svgs/mentor";
+import { MentoringApplyStatus, MentoringApprovalStatus } from "@/types/mentor";
 import { convertISODateToKoreanTime } from "@/utils/datetimeUtils";
 import useExpandCardClickHandler from "./hooks/useExpandCardClickHandler";
 
@@ -33,7 +34,23 @@ const MentorExpandChatCard = ({
     mentoringId,
     initChecked: isChecked,
   });
-  const { handleAccept, handleReject } = usePatchApprovalStatusHandler();
+  const { mutate: patchApprovalStatus } = usePatchApprovalStatus();
+
+  const handleAccept = (targetMentoringId: number) => {
+    patchApprovalStatus({ status: MentoringApprovalStatus.APPROVED, mentoringId: targetMentoringId });
+  };
+
+  const handleReject = async (targetMentoringId: number) => {
+    const ok = await customConfirm({
+      title: "멘티 신청을 정말 거절하시겠어요?",
+      content: "누적 5회 거절 시 멘토 활동에 제약이 있을 수 있어요. 다시 한 번 고민해보세요.",
+      icon: IconUnSmile,
+      approveMessage: "다시 생각해볼게요",
+      rejectMessage: "그래도 거절할래요",
+    });
+    if (ok) return;
+    patchApprovalStatus({ status: MentoringApprovalStatus.REJECTED, mentoringId: targetMentoringId });
+  };
 
   return (
     <div className="w-full overflow-hidden border-b border-k-50" ref={lastElementRef}>
